[feat#1] 소셜 로그인 구현

See merge request s13-webmobile1-sub1/S13P11E104!155

+++ Backend/src/main/java/com/e104/reciplay/repository/SubSubscriptionRepository.java
@@ -4,14 +4,13 @@ import com.e104.reciplay.entity.Subscription;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import javax.swing.text.html.Option;
 import java.util.List;
 import java.util.Optional;
 
 @Repository
 public interface SubSubscriptionRepository extends JpaRepository<Subscription, Long> {
     List<Subscription> findByUserId(Long userId);
-    Long countByInstructorId(Long instructorId);
+    Integer countByInstructorId(Long instructorId);
     Optional<Subscription> findByInstructorIdAndUserId(Long instructorId, Long userId);
     void deleteByInstructorIdAndUserId(Long instructorId, Long userId);
 }
